# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SkyRoute Flight Booking API",
    "version": "1.0.0",
    "description": "Comprehensive REST API documentation for SkyRoute. Includes user authentication, flight schedules & searches, seat allocations, booking workflows, and administrative management."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server Instance"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session token returned by register or login endpoints. Pass as `Authorization: Bearer <token>` or as a cookie (`skyroute_session`)."
      }
    },
...
  },
  "paths": {
...
    "/api/bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create Booking",
        "description": "Create a new flight reservation with passenger details and payment. Supports guest and authenticated bookings.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "flightId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 80
                        },
                        "cabin": {
                          "type": "string",
                          "enum": [
                            "economy",
                            "business",
                            "first"
                          ]
                        },
                        "seatIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "nullable": true,
                            "pattern": "^\\d{1,3}[A-K]$"
                          },
                          "maxItems": 9
                        }
                      },
                      "required": [
                        "flightId",
                        "cabin",
                        "seatIds"
                      ]
                    },
                    "minItems": 1,
                    "maxItems": 6
                  },
                  "tripType": {
                    "type": "string",
                    "enum": [
                      "one-way",
                      "round-trip",
                      "multi-city"
                    ]
                  },
                  "contactEmail": {
                    "type": "string",
                    "maxLength": 120,
                    "format": "email"
                  },
                  "contactPhone": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 20
                  },
                  "passengers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string",
                          "enum": [
                            "Mr",
                            "Mrs",
                            "Miss",
                            "Ms",
                            "Dr"
                          ]
                        },
                        "firstName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 50
                        },
                        "lastName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 50
                        },
                        "dateOfBirth": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "gender": {
                          "type": "string",
                          "enum": [
                            "male",
                            "female"
                          ]
                        },
                        "passportNumber": {
                          "type": "string",
                          "maxLength": 20
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "adult",
                            "child",
                            "infant"
                          ]
                        }
                      },
                      "required": [
                        "title",
                        "firstName",
                        "lastName",
                        "dateOfBirth",
                        "gender",
                        "type"
                      ]
                    },
                    "minItems": 1,
                    "maxItems": 9
                  },
                  "payment": {
                    "type": "object",
                    "properties": {
                      "method": {
                        "type": "string",
                        "enum": [
                          "card",
                          "transfer",
                          "wallet"
                        ]
                      },
                      "cardHolder": {
                        "type": "string",
                        "maxLength": 80
                      },
                      "cardNumber": {
                        "type": "string",
                        "maxLength": 25
                      },
                      "expiry": {
                        "type": "string",
                        "maxLength": 7
                      },
                      "cvv": {
                        "type": "string",
                        "maxLength": 4
                      },
                      "senderName": {
                        "type": "string",
                        "maxLength": 80
                      },
                      "forceFailure": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "method"
                    ]
                  }
                },
                "required": [
                  "legs",
                  "contactEmail",
                  "contactPhone",
                  "passengers",
                  "payment"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking confirmed with PNR reference."
          },
          "400": {
            "description": "Validation error, seat unavailable, or invalid payment."
          }
        }
      },
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "List User Bookings",
        "description": "List all bookings belonging to the signed-in account.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user bookings."
          },
          "401": {
            "description": "Authentication required."
          }
        }
      }
    },
    "/api/bookings/lookup": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Guest Booking Lookup",
        "description": "Retrieve booking details using 6-character PNR reference code and passenger surname.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pnr": {
                    "type": "string",
                    "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{6}$"
                  },
                  "surname": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  }
                },
                "required": [
                  "pnr",
                  "surname"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking details and cancellation quote."
          },
          "404": {
            "description": "Booking reference or surname does not match."
          }
        }
      }
    },
    "/api/bookings/{pnr}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Booking by PNR",
        "description": "Retrieve booking details by PNR.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{6}$"
            },
            "required": true,
            "name": "pnr",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details and cancellation quote."
          },
          "404": {
            "description": "Booking reference not found."
          }
        }
      }
    },
    "/api/bookings/{pnr}/cancel": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Cancel Booking",
        "description": "Cancel a booking and calculate refund according to policy. Accepts account session or surname verification.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{6}$"
            },
            "required": true,
            "name": "pnr",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "surname": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking cancelled with refund receipt."
          },
          "400": {
            "description": "Surname missing for guest cancellation or flight already departed."
          },
          "404": {
            "description": "Booking not found."
          }
        }
      }
    },
...
    "/api/admin/bookings": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin List Bookings",
        "description": "Paginated list of all bookings across all customers.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "confirmed",
                "cancelled",
                "pending"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of bookings."
          },
          "403": {
            "description": "Admin privileges required."
          }
        }
      }
    },
    "/api/admin/bookings/{pnr}/cancel": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Cancel Booking",
        "description": "Administrative override to cancel any booking.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{6}$"
            },
            "required": true,
            "name": "pnr",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking cancelled."
          },
          "403": {
            "description": "Admin privileges required."
          }
        }
      }
    },
...
  }
}